# Improve performance of search boundries for int64 column