--- a/src/FlowtideDotNet.Core/ColumnStore/DataColumns/Int64Column.cs
+++ b/src/FlowtideDotNet.Core/ColumnStore/DataColumns/Int64Column.cs
@@ -132,9 +132,9 @@ namespace FlowtideDotNet.Core.ColumnStore
             var val = dataValue.AsLong;
             if (desc)
             {
-                return BoundarySearch.SearchBoundries(_data, val, start, end, Int64ComparerDesc.Instance);
+                return BoundarySearch.SearchBoundriesInt64Desc(_data, start, end, val);
             }
-            return BoundarySearch.SearchBoundries(_data, val, start, end, Int64Comparer.Instance);
+            return BoundarySearch.SearchBoundriesInt64Asc(_data, start, end, val);
         }
 
         public int Update(in int index, in IDataValue value)