--- a/src/FlowtideDotNet.Core/FlowtideDotNet.Core.csproj
+++ b/src/FlowtideDotNet.Core/FlowtideDotNet.Core.csproj
@@ -18,6 +18,9 @@
   <ItemGroup>
     <PackageReference Include="Apache.Arrow" Version="16.1.0" />
     <PackageReference Include="System.IO.Hashing" Version="8.0.0" />
+
+	  <PackageReference Include="Fody" Version="6.6.3" PrivateAssets="all" />
+	  <PackageReference Include="InlineIL.Fody" Version="1.7.2" PrivateAssets="all" />
   </ItemGroup>
 
   <ItemGroup>