--- a/tests/FlowtideDotNet.Benchmarks/BoundarySearchInt64Benchmarks.cs
+++ b/tests/FlowtideDotNet.Benchmarks/BoundarySearchInt64Benchmarks.cs
@@ -1,28 +1,26 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Numerics;
-using System.Runtime.CompilerServices;
-using System.Runtime.Intrinsics;
-using System.Runtime.Intrinsics.X86;
-using System.Text;
-using System.Threading.Tasks;
+﻿//// Licensed under the Apache License, Version 2.0 (the "License")
+//// you may not use this file except in compliance with the License.
+//// You may obtain a copy of the License at
+////
+////     http://www.apache.org/licenses/LICENSE-2.0
+////  
+//// Unless required by applicable law or agreed to in writing, software
+//// distributed under the License is distributed on an "AS IS" BASIS,
+//// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//// See the License for the specific language governing permissions and
+//// limitations under the License.
+
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Diagnosers;
 using FlowtideDotNet.Core.ColumnStore;
 using FlowtideDotNet.Core.ColumnStore.TreeStorage;
 using FlowtideDotNet.Core.ColumnStore.Utils;
 using FlowtideDotNet.Storage.Memory;
-using InlineIL;
-using static InlineIL.IL.Emit;
-using static Substrait.Protobuf.Expression.Types.Literal.Types;
 
 namespace FlowtideDotNet.Benchmarks
 {
     [HardwareCounters(HardwareCounter.BranchMispredictions, HardwareCounter.BranchInstructions)]
-    public class BinarySearchBenchmarks
+    public class BoundarySearchInt64Benchmarks
     {
         private const int SearchSpace = 1024 * 1;