--- a/tests/FlowtideDotNet.Benchmarks/FlowtideDotNet.Benchmarks.csproj
+++ b/tests/FlowtideDotNet.Benchmarks/FlowtideDotNet.Benchmarks.csproj
@@ -10,6 +10,7 @@
 
   <ItemGroup>
     <PackageReference Include="BenchmarkDotNet" Version="0.13.12" />
+    <PackageReference Include="BenchmarkDotNet.Diagnostics.Windows" Version="0.13.12" />
   </ItemGroup>
 
   <ItemGroup>