Benchmark:

| Method                                | Mean     | Error    | StdDev   | BranchInstructions/Op | BranchMispredictions/Op |
|-------------------------------------- |---------:|---------:|---------:|----------------------:|------------------------:|
| ArrayBinarySearchBenchmark            | 37.28 us | 0.452 us | 0.353 us |                73,647 |                   3,058 |
| ColumnBoundarySearchBenchmark         | 24.04 us | 0.149 us | 0.139 us |                44,430 |                      26 |
| NativeLongListBoundarySearchBenchmark | 14.56 us | 0.042 us | 0.039 us |                12,489 |                       9 |
| Int64ColumnBoundarySearchBenchmark    | 16.33 us | 0.151 us | 0.141 us |                17,667 |                      12 |